Remove ActiveOffersTab folder, move logic to ActiveOffersTable
