--- src/pages/OffersPage/components/ActiveOffersTab/ActiveOffersTab.module.less
@@ -1,8 +1,0 @@
-.rootTable :global {
-  .ant-table-tbody > tr > td {
-    width: 210px;
-  }
-  .ant-table-tbody > tr > td:last-child {
-    width: 180px;
-  }
-}
