--- src/pages/OffersPage/components/ActiveOffersTab/index.ts
@@ -1,1 +1,0 @@
-export * from './ActiveOffersTab'
