+++ src/pages/OffersPage/components/ActiveOffersTable/ActiveOffersTable.module.less
@@ -1,3 +1,12 @@
+.rootTable :global {
+  .ant-table-tbody > tr > td {
+    width: 210px;
+  }
+  .ant-table-tbody > tr > td:last-child {
+    width: 180px;
+  }
+}
+
 .actionsButtons {
   display: flex;
   align-items: center;
