+++ src/pages/OffersPage/components/ActiveOffersTable/constants.ts
