+++ src/pages/OffersPage/components/ActiveOffersTable/hooks/useLenderLoansAndOffers.ts
@@ -1,18 +1,11 @@
-import { useMemo, useState } from 'react'
+import { useMemo } from 'react'
 
 import { useWallet } from '@solana/wallet-adapter-react'
 import { useQuery } from '@tanstack/react-query'
 import { produce } from 'immer'
-import { first, groupBy, map, sumBy } from 'lodash'
 import { create } from 'zustand'
 
-import { SearchSelectProps } from '@banx/components/SearchSelect'
-import { SortOption } from '@banx/components/SortDropdown'
-import { createSolValueJSX } from '@banx/components/TableComponents'
-
 import { Offer, fetchLenderLoansAndOffers } from '@banx/api/core'
-
-import { DEFAULT_SORT_OPTION } from './constants'
 
 interface HiddenNftsAndOffersState {
   mints: string[]
@@ -96,59 +89,5 @@ export const useLenderLoansAndOffers = () => {
     loans,
     offers: data?.offers ?? {},
     loading: isLoading,
-  }
-}
-
-interface SearchSelectOption {
-  collectionName: string
-  collectionImage: string
-}
-
-export const useActiveOffersTab = () => {
-  const { loans, loading } = useLenderLoansAndOffers()
-
-  const [selectedOptions, setSelectedOptions] = useState<string[]>([])
-  const [sortOption, setSortOption] = useState<SortOption>(DEFAULT_SORT_OPTION)
-
-  const searchSelectOptions = useMemo(() => {
-    const loansGroupedByCollection = groupBy(loans, ({ nft }) => nft.meta.collectionName)
-
-    return map(loansGroupedByCollection, (groupedLoan) => {
-      const firstLoanInGroup = first(groupedLoan)
-      const { collectionName = '', collectionImage = '' } = firstLoanInGroup?.nft.meta || {}
-      const taken = sumBy(groupedLoan, (nft) => nft.bondTradeTransaction.solAmount)
-
-      return { collectionName, collectionImage, taken }
-    })
-  }, [loans])
-
-  const searchSelectParams: SearchSelectProps<SearchSelectOption> = {
-    options: searchSelectOptions,
-    optionKeys: {
-      labelKey: 'collectionName',
-      valueKey: 'collectionName',
-      imageKey: 'collectionImage',
-      secondLabel: {
-        key: 'taken',
-        format: (value: number) => createSolValueJSX(value, 1e9),
-      },
-    },
-    selectedOptions,
-    labels: ['Collection', 'Taken'],
-    onChange: setSelectedOptions,
-  }
-
-  const sortParams = {
-    option: sortOption,
-    onChange: setSortOption,
-  }
-
-  return {
-    loans,
-    loading,
-    sortViewParams: {
-      searchSelectParams,
-      sortParams,
-    },
   }
 }
