+++ src/pages/OffersPage/components/OffersPageContent/OffersPageContent.tsx
@@ -1,7 +1,7 @@
 import { Tabs, useTabs } from '@banx/components/Tabs'
 
 import { OFFERS_TABS, OffersTabName } from '../../constants'
-import ActiveOffersTab from '../ActiveOffersTab/ActiveOffersTab'
+import ActiveOffersTable from '../ActiveOffersTable'
 import PendingOffersTab from '../PendingOffersTab'
 
 import styles from './OffersPageContent.module.less'
@@ -18,7 +18,7 @@ const OffersPageContent = () => {
       <Tabs className={styles.tabs} tabs={offersTabs} value={tabValue} setValue={setTabValue} />
       <div className={styles.tabContent}>
         {tabValue === OffersTabName.PENDING && <PendingOffersTab />}
-        {tabValue === OffersTabName.ACTIVE && <ActiveOffersTab />}
+        {tabValue === OffersTabName.ACTIVE && <ActiveOffersTable />}
       </div>
     </div>
   )
